+++ pkg/cache/inmemory_test.go
@@ -5,20 +5,20 @@ package cache
 
 import (
 	"context"
+
+	"github.com/go-kit/kit/log"
+
 	"testing"
 	"time"
 
-	"github.com/go-kit/kit/log"
-	"github.com/pkg/errors"
 	prom_testutil "github.com/prometheus/client_golang/prometheus/testutil"
-
 	"github.com/thanos-io/thanos/pkg/testutil"
 )
 
-func TestMemcachedCache(t *testing.T) {
+func TestInmemoryCache(t *testing.T) {
 	t.Parallel()
 
-	// Init some data to conveniently define test cases later one.
+	// Init some data to conveniently define test cases later on.
 	key1 := "key1"
 	key2 := "key2"
 	key3 := "key3"
@@ -28,12 +28,15 @@ func TestMemcachedCache(t *testing.T) {
 
 	tests := map[string]struct {
 		setup        map[string][]byte
+		ttl          time.Duration
+		testTTL      bool
 		mockedErr    error
 		fetchKeys    []string
 		expectedHits map[string][]byte
 	}{
 		"should return no hits on empty cache": {
 			setup:        nil,
+			ttl:          time.Duration(time.Hour),
 			fetchKeys:    []string{key1, key2},
 			expectedHits: map[string][]byte{},
 		},
@@ -43,39 +46,51 @@ func TestMemcachedCache(t *testing.T) {
 				key2: value2,
 				key3: value3,
 			},
+			ttl:       time.Duration(time.Hour),
 			fetchKeys: []string{key1},
 			expectedHits: map[string][]byte{
 				key1: value1,
 			},
+		},
+		"should return a miss due to TTL": {
+			setup: map[string][]byte{
+				key1: value1,
+				key2: value2,
+				key3: value3,
+			},
+			// setting ttl for the values as 250 ms
+			ttl:          time.Duration(250 * time.Millisecond),
+			testTTL:      true,
+			fetchKeys:    []string{key1},
+			expectedHits: map[string][]byte{},
 		},
 		"should return hits and misses on partial hits": {
 			setup: map[string][]byte{
 				key1: value1,
 				key2: value2,
 			},
+			ttl:          time.Duration(time.Hour),
 			fetchKeys:    []string{key1, key3},
 			expectedHits: map[string][]byte{key1: value1},
 		},
-		"should return no hits on memcached error": {
-			setup: map[string][]byte{
-				key1: value1,
-				key2: value2,
-				key3: value3,
-			},
-			mockedErr:    errors.New("mocked error"),
-			fetchKeys:    []string{key1},
-			expectedHits: nil,
-		},
 	}
-
 	for testName, testData := range tests {
 		t.Run(testName, func(t *testing.T) {
-			memcached := newMockedMemcachedClient(testData.mockedErr)
-			c := NewMemcachedCache("test", log.NewNopLogger(), memcached, nil)
+			conf := []byte(`
+max_size: 1MB
+max_item_size: 2KB
+`)
+			c, _ := NewInMemoryCache("test", log.NewNopLogger(), nil, conf)
 
 			// Store the postings expected before running the test.
 			ctx := context.Background()
-			c.Store(ctx, testData.setup, time.Hour)
+			c.Store(ctx, testData.setup, testData.ttl)
+
+			// Add delay to test expiry functionality.
+			if testData.testTTL {
+				// The delay will be based on the TTL itself.
+				time.Sleep(testData.ttl)
+			}
 
 			// Fetch postings from cached and assert on it.
 			hits := c.Fetch(ctx, testData.fetchKeys)
@@ -88,41 +103,13 @@ func TestMemcachedCache(t *testing.T) {
 	}
 }
 
-// mockedMemcachedClient is a mocked memcached client for testing.
-type mockedMemcachedClient struct {
-	cache       map[string][]byte
-	getMultiErr error
-}
-
-// newMockedMemcachedClient returns a mocked memcached client.
-func newMockedMemcachedClient(getMultiErr error) *mockedMemcachedClient {
-	return &mockedMemcachedClient{
-		cache:       map[string][]byte{},
-		getMultiErr: getMultiErr,
-	}
-}
-
-func (c *mockedMemcachedClient) GetMulti(_ context.Context, keys []string) map[string][]byte {
-	if c.getMultiErr != nil {
-		return nil
-	}
-
-	hits := map[string][]byte{}
-
-	for _, key := range keys {
-		if value, ok := c.cache[key]; ok {
-			hits[key] = value
-		}
-	}
-
-	return hits
-}
-
-func (c *mockedMemcachedClient) SetAsync(_ context.Context, key string, value []byte, _ time.Duration) error {
-	c.cache[key] = value
-	return nil
-}
-
-func (c *mockedMemcachedClient) Stop() {
-	// Nothing to do.
+func TestNewInMemoryCache(t *testing.T) {
+	// Should return error when the max size of the cache is smaller than the max size of
+	conf := []byte(`
+max_size: 2KB
+max_item_size: 1MB
+`)
+	cache, err := NewInMemoryCache("test", log.NewNopLogger(), nil, conf)
+	testutil.NotOk(t, err)
+	testutil.Equals(t, (*InMemoryCache)(nil), cache)
 }